+++ libs/general.py
@@ -1,0 +1,9 @@
+import shutil
+
+def copyDirectory(src, dest):
+    try:
+        shutil.copytree(src, dest)
+    except shutil.Error as e:
+        print('Directory not copied. Error: %s' % e)
+    except OSError as e:
+        print('Directory not copied. Error: %s' % e)